Refactor ItemList and Item, add AddItemForm component
